refactor: move auth logic out of server.js into controllers
